requirements.txt
blobfish-ejbca-client-python==0.1.0
